out/production/multi-agent-legal-review-system/requirements.txt
# === 기존 의존성 (규칙 기반 엔진) ===
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
httpx>=0.27.0
python-dotenv>=1.0.0

# === LangGraph 엔진 의존성 (USE_LANGGRAPH=true 시 필요) ===
google-genai>=1.0.0
langgraph>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
